Refactor property config and RACI helpers into modules
